Move all error handlers to one place
